Processor has no dependencies to simplify setup in pom.xml
