+++ config/metadata-processor/src/test/java/io/helidon/config/metadata/processor/ConfigMetadataHandlerTest.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2021 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.helidon.config.metadata.processor;
+
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class ConfigMetadataHandlerTest {
+    @Test
+    void testToConfigKey() {
+        ConfigMetadataHandler p = new ConfigMetadataHandler();
+        assertAll(
+                () -> assertThat(p.toConfigKey("maxInitialLineLength"), is("max-initial-line-length")),
+                () -> assertThat(p.toConfigKey("port"), is("port")),
+                () -> assertThat(p.toConfigKey("listenAddress"), is("listen-address"))
+        );
+    }
+}
