+++ pom.xml
@@ -210,6 +210,7 @@
                         <source>11</source>
                         <target>11</target>
                         <release>11</release>
+                        <forceJavacCompilerUse>true</forceJavacCompilerUse>
                         <compilerArgs>
                             <arg>-Xlint:unchecked</arg>
                             <!--
